revert to dual state

+++ src/js/components/Toast.js
@@ -4,13 +4,19 @@ import { CSSTransition, TransitionGroup } from 'react-transition-group';
 
 export default function Toast(props) {
   const { toastList, setToastList, autoDelete } = props;
+  const [list, setList] = React.useState(toastList);
+
+  React.useEffect(() => {
+    setList(toastList);
+  }, [toastList]);
 
   React.useEffect(() => {
     const interval = setInterval(() => {
-      if (autoDelete && toastList.length) {
+      console.log('int');
+      if (autoDelete && toastList.length && list.length) {
         deleteToast(toastList[0].id);
       }
-    }, 1000);
+    }, 3000);
 
     return () => {
       clearInterval(interval);
@@ -18,32 +24,27 @@ export default function Toast(props) {
   }, [toastList, autoDelete]);
 
   function deleteToast(id) {
-    const toastListItem = toastList.findIndex((e) => e.id === id);
-    toastList.splice(toastListItem, 1);
-    setToastList([...toastList]);
+    setList((items) => items.filter((item) => item.id !== id));
+    setToastList((items) => items.filter((item) => item.id !== id));
   }
 
   return (
     <TransitionGroup className={`notification-container `}>
-      {toastList.map((toast, i) => {
+      {list.map(({ id, title, description }) => {
         return (
-          <CSSTransition
-            key={toast.id}
-            timeout={200}
-            classNames="toast-transition"
-          >
-            <div className={`notification2 toast toast--${toast.type}`}>
+          <CSSTransition key={id} timeout={500} classNames="toast-transition">
+            <div className={`notification2 toast`}>
               <button
                 onClick={() => {
-                  deleteToast(toast.id);
+                  deleteToast(id);
                 }}
               >
                 X
               </button>
 
               <div>
-                <p className="notification-title">{toast.title}</p>
-                <p className="notification-message">{toast.description}</p>
+                <p className="notification-title">{title}</p>
+                <p className="notification-message">{description}</p>
               </div>
             </div>
           </CSSTransition>
